+++ changelog.md
@@ -12,6 +12,8 @@ Codion Change Log
 ### is.codion.common.rmi
 - SerializationWhitelist, now allows arrays but checks the component type.
 - SerializationWhitelist.DryRun, now excludes arrays and includes the component type.
+### is.codion.swing.common.ui.tools
+- LoadTestPanel.loadTestPanel() factory method added, constructor now private.
 ### is.codion.swing.framework.model.tools
 - DomainToString bug fixed, now handles the mapping from an unknown column data type to an object based Column correctly, DatabaseDomain some renaming.
 - DomainGeneratorModel, table type column moved.